Clients get SOME of each other's messages
I'm not sure why, but sometimes, messages are not properly updated for
each client. The problem probably regards accessing the users table (to
check the lastupdate column) twice at the same time, but I had thought
that JavaScript was technically single-threaded always, which is why I'm
not sure. However, since this is mere experimentation, I may not try to
fix the problem, since I already achieved my goal of having clients
communicate.